In a stateless model, the state of execution focuses around a "registry" maintained by the SCXMLExecutor, which performs book-keeping functions concerning a particular execution instance of the state machine.
The NotificationRegistry and SCXMLExecutor have to deal with the absence of the Observable interface in the Commons SCXML object model, and therefore, what was earlier effectively:

 addListener(Observable, SCXMLListener)

now becomes (effectively):

 * addListener(SCXML, SCXMLListener)
 * addListener(TransitionTarget, SCXMLListener)
 * addListener(Transition, SCXMLListener)

since listeners may be attached to o.a.c.s.m.{SCXML,TransitionTarget,Transition} which are the interesting elements for tracking execution.

Some qualifiers and access modifiers have been appropriately constrained in this commit.



git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/scxml/branches/STATELESS_MODEL@373328 13f79535-47bb-0310-9956-ffa450edef68