One of the changes between the July 05 WD [1] (which we're currently close to par with) and the January 06 WD [2] is that transition targets can no longer be anonymous child states of <transition>.
Furthermore, the <target> child of <transition> has been changed into an attribute. This definitely makes the document more compact, since snippets like this:

<transition event="foo.bar">
  <target next="foobarState"/>
</transition>

now turn into:

<transition event="foo.bar" target="foobarState"/>

However, anonymous target states are no longer mentioned in the WD, so the following construct (though not yet removed from the implementation), is not specified by the latest WD, and shouldn't be used:

<transition event="foo.bar">
 <target>
  <state>
   <!-- content of anonymous state -->
  </state>
 </target>
</transition>

So another todo (now this gets the privilege of being low priority ;-) will be to redo documents in the RDC and Shale usecases to sync up. This is low priority since the Commons SCXML implementation still continues to support either case. However, as we tighten up the implementation against the WD, we will probably shed that support (and adjust the usecases accordingly).

[1] http://www.w3.org/TR/2005/WD-scxml-20050705/
[2] http://www.w3.org/TR/2006/WD-scxml-20060124/


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/scxml/trunk@382958 13f79535-47bb-0310-9956-ffa450edef68