Partial implementation for the datamodel section [1] (Section 5) of the latest W3C SCXML WD [2].
A <datamodel> tag may contain any number of <data> elements. Each data element can contain an XML data tree (a degenerate flavor exists matching the <var> element functionality as well). The XML data trees may be fetched from an external source (not yet implemented) or inline (implemented). The datamodel may be "distributed" within the document per author's convenience, that is, it may be a child of the <scxml> element or the <state> element. Visibility of data is structured like blocks in a procedural language (nested blocks -> nested states). A builtin function Data() is available to examine the Nodes within a XML data tree (for use with JEXL and EL expressions). Accompanied by a starter set of test cases and a couple of sample documents for illustration.

Also aligning SCXML Context to JEXL Context (provide a getVars() instead of a iterator()).

[1] http://www.w3.org/TR/scxml/#data
[2] http://www.w3.org/TR/2006/WD-scxml-20060124/


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/scxml/trunk@388312 13f79535-47bb-0310-9956-ffa450edef68