Implemented <invoke> (Section 4.4) and its children <param> (Section 4.5) and <finalize> (Section 4.6) -- all sections from the January 06 W3C WD [ http://www.w3.org/TR/scxml/ ]. With this addition, Commons SCXML aligns with the Jan '06 WD.
The primary addition is the Invoker interface, which is used to define the possible interactions between the parent state machine (executor) and the types of invocable activities. Invocable activities must first register an Invoker implementation class for the appropriate "targettype" (attribute of <invoke>) with the parent SCXMLExecutor.

The communication link between the parent state machine executor and the invoked activity is a bi-directional events pipe. All events triggered on the parent state machine get forwarded to the invoked activity. The processing semantics for these events depend upon the "targettype", and thereby vary per concrete implementation of the Invoker interface.

The Invoker "lifecycle" is outlined below: 

 * Instantiation via Class.newInstance() (Invoker implementation requires accessible constructor).
 * Configuration (setters for parent state ID and SCInstance).
 * Initiation of invoked activity via invoke() method, passing the source URI and the map of params.
 * Zero or more bi-directional event triggers.
 * Either completion or cancellation.

These additions introduce backwards incompatible changes to the SCXMLSemantics interface (two new methods).

I will bring up the couple of open issues (such as event names TBD in WD and handling invocation failure) at the related W3C face to face meeting later this week.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/scxml/trunk@415512 13f79535-47bb-0310-9956-ffa450edef68