The Commons SCXML object model is now Serializable. Includes some non-functional changes such as reorganization of imports.
To demonstrate serializability, the test suite has been altered such that for each relevant test involving SCXML document IO, the following sequence is followed:

a) SCXMLDigester parses the document
b) The parsed result is serialized
c) The serialized form is read back into an SCXML object

The result in (c) is now used to run the tests, as against the result in (a) earlier. The build leaves behind the serialized objects in the "target/serialization" directory (and these get cleaned up on a clean build).

SCXML-29


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/scxml/trunk@452661 13f79535-47bb-0310-9956-ffa450edef68