The February '07 SCXML WD [1] makes some changes to the SCXML vocabulary, particularly in the context of the <parallel> element (which is now a orthogonal state instead of qualifying the enclosing state to be orthogonal) and built-in actions (<var>, <cancel>, <exit> have been removed in favor of a more open content model for actions -- we will continue to support them in a Commons SCXML namespace since they are useful).
These changes require updates to the Commons SCXML parser, and while it will be possible to retain the original parser and make it WD version sensitive, I dislike the original class and method names which are implementation specific. The new parser will be a class of its own (with a better name), and the original parser will be deprecated in v0.7 (though retained till v1.0 for compatibility with v0.x releases).

[1] http://www.w3.org/TR/2007/WD-scxml-20070221/


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/scxml/trunk@521501 13f79535-47bb-0310-9956-ffa450edef68