SCXML-45 Payload of events sent to current scxml session using <send> tag not injected into engine
This has been implemented (with a test case), however note the following caveat --

The spec doesn't clarify how multiple <send> elements that create derived events should be handled, so for example:

<onentry>
  <send event="ev.foo" namelist="alpha beta"/>
  <send event="ev.bar" namelist="gamma delta"/>
</onentry>

I think they should be processed together (this makes sense to leverage parallel regions for example), and due to that '_eventdata' becomes ambiguous in this scenario. The Commons SCXML implementation introduces an implicit variable '_eventdatamap' for such scenarios wherein the event datas are stored keyed by event name.

So, the two <send> events above could be processed by two regions like so:

<parallel>

  <state id="region1">

     <transition event="ev.foo" cond="_eventdatamap['ev.foo'].alpha eq 'somevalue'"
                 target="..." />

     <!-- ... -->

  </state>

  <state id="region2">

     <transition event="ev.bar" cond="_eventdatamap['ev.bar'].delta eq 'othervalue'"
                 target="..." />

     <!-- ... -->

  </state>

  <!-- ... -->

</parallel>

To summarize, the _eventdatamap variable needs to be used in association with "derived" (such as <send> being discussed here) events. Also note that this behavior may change if there is clarity in the specification at some point.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/scxml/trunk@542582 13f79535-47bb-0310-9956-ffa450edef68